// server/apps/api/src/routes/order/order.model.ts
export const GetManageOrderListQuerySchema = PaginationQuerySchema.extend({
  status: OrderStatusSchema.optional(),
  startDate: z.string().optional(),
  endDate: z.string().optional(),
  customerName: z.string().optional(),
  orderCode: z.string().optional()
})

export const GetOrderListQuerySchema = PaginationQuerySchema.extend({
  status: OrderStatusSchema.optional()
})

export const GetOrderDetailResSchema = z.object({
  message: z.string().optional(),
  data: OrderSchema.extend({
    items: z.array(ProductSKUSnapshotSchema),
    orderCode: z.string().nullable().optional()
  })
})

export const GetManageOrderDetailResSchema = z.object({
  message: z.string().optional(),
  data: OrderSchema.extend({
    items: z.array(ProductSKUSnapshotSchema),
    orderCode: z.string().nullable().optional()
  })
})
// ...
export type GetManageOrderListQueryType = z.infer<typeof GetManageOrderListQuerySchema>
export type GetManageOrderListResType = z.infer<typeof GetManageOrderListResSchema>
export type GetManageOrderDetailResType = z.infer<typeof GetManageOrderDetailResSchema>